refactor(landmarks): split chrome.tsx into chrome/ folder
Extract the 1728-line chrome.tsx into a chrome/ directory with
dedicated modules for sections, primitives, genes-ternary math,
genes-legend, topbar, layers-panel, control-panel, and mobile-chrome.

- Move blendHex/parseHexRgb to helpers.ts (generic color math)
- Remove deprecated InspectPanel alias
- Barrel chrome/index.ts preserves existing imports
- All data-testid attributes unchanged
- tsc --noEmit + build pass
