+++ frontend/src/widgets/landmarks/chrome/index.ts
@@ -1,0 +1,5 @@
+export { Topbar } from "./topbar";
+export { LayersPanel } from "./layers-panel";
+export { ControlPanel } from "./control-panel";
+export { MobileSectionChrome } from "./mobile-chrome";
+export type { PanelSectionId } from "./sections";
